+++ TDS_Common/Dto/SyncedPlayerSettingsDto.cs
@@ -2,7 +2,7 @@
 {
     public class SyncedPlayerSettingsDto
     {
-        public byte Language;
+        public short Language;
         public bool Hitsound;
         public bool Bloodscreen;
         public bool FloatingDamageInfo;
