+++ TDS_Common/Dto/SyncedSettingsDto.cs
@@ -2,8 +2,8 @@
 {
     public class SyncedServerSettingsDto
     {
-        public int DistanceToSpotToPlant;
-        public int DistanceToSpotToDefuse;
+        public float DistanceToSpotToPlant;
+        public float DistanceToSpotToDefuse;
         public int RoundEndTime;
         public int MapChooseTime;
         public int TeamOrderCooldownMs;
